% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSummary.R
\name{calcSummary}
\alias{calcSummary}
\title{calcSummary function creates columns for good to bad ratio and net score for each week for each category.}
\usage{
calcSummary(sq_data)
}
\description{
This function calculates some summaries and adds them as columns to this dataframe
}
\examples{
summary_df <- calcSummary(sq_data)
}
\keyword{quantified}
\keyword{self}
