% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSqTemplate.R
\name{createSqTemplate}
\alias{createSqTemplate}
\title{createSqTemplate function creates a new google sheet in the template form which we follow for the quantified self data recording process}
\usage{
createSqTemplate(title = "Self Quant", worksheet = "Template")
}
\arguments{
\item{title}{parameter for the name of the google sheet, defaults to Self Quant}

\item{workSheet}{parameter for the name of the worksheet within the doc, defaults to Template}
}
\description{
This function creates a new google sheet called "Self Quant", and worksheet called "Template"
}
\examples{
some example here
}
\keyword{quantified}
\keyword{self}
