% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuant.R
\name{getQuant}
\alias{getQuant}
\title{getQuant function reads in your weekly stats from your google steet}
\usage{
getQuant(title = "Self Quant", workSheet = "Template")
}
\arguments{
\item{title}{parameter for the name of the google sheet, defaults to Self Quant}

\item{workSheet}{parameter for the name of the worksheet within the doc, defaults to Template}
}
\description{
This function reads your monthly stats from a google sheet to a dataframe
}
\examples{
some example here
}
\keyword{quantified}
\keyword{self}
