% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFlower.R
\name{plotFlower}
\alias{plotFlower}
\title{plotFlower plots your good to bad ratio score for each metric over the past 4 weeks}
\usage{
plotFlower(sq_summary)
}
\arguments{
\item{summary}{parameter for the summarised dataset created using the calcSummary() function}
}
\description{
This function looks at variation in your good to bad ratio for each metric over the last 4 weeks. Each petal of the flower is one of the metrics.
The score in the middle represents the average good to bad ratio score across all metrics. If it is greater than 1, then overall you have more good, and if it's smaller than 1, you have more bad.
}
\examples{
plotFlower(summary = summary_df)

}
\keyword{quantified}
\keyword{self}
