% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotWeek.R
\name{plotWeek}
\alias{plotWeek}
\title{plotWeek plots your net score each week in the past 4 weeks}
\usage{
plotWeek(sq_summary)
}
\arguments{
\item{summary}{parameter for the summarised dataset created using the calcSummary() function}
}
\description{
This function looks at variation in your net score for each metric over the last 4 weeks
}
\examples{
plotWeek(summary = sq_data)

}
\keyword{quantified}
\keyword{self}
