% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQuant.R
\name{getQuant}
\alias{getQuant}
\title{getQuant function reads in your weekly stats from your google steet}
\usage{
getQuant(title = "Self Quant", workSheet = "Template")
}
\arguments{
\item{title}{parameter for the name of the google sheet, defaults to Self Quant}

\item{workSheet}{parameter for the name of the worksheet within the doc, defaults to Template}
}
\description{
This function reads your monthly stats from a google sheet to a dataframe
It also calculates some summaries in additional columns,
These are the net scores for each week (good scores - bad scores) and the good to bad ratio (good scores/ bad scores)
}
\examples{
sq_data <- getQuant()

So if you have this sheet in your google sheets:
https://docs.google.com/spreadsheets/d/1ulmrOuD2_jCYJ4TxTVcNZCPqUVR5wxx6DoXW_uz-6lk/edit?usp=sharing
You can read this in with:
sq_data <- getQuant(title="Self Quant", workSheet = "Template")
}
\keyword{quantified}
\keyword{self}
